avg weight fix

+++ src/utils/settings.h
@@ -63,7 +63,7 @@ class Settings
 		unsigned int info_totalAdamDistance;
 
 		unsigned int info_totalBodyparts;
-		unsigned int info_totalWeight;
+		float info_totalWeight;
 
 		int *winWidth;
 		int *winHeight;
